ITensorNetworks

Documentation for ITensorNetworks.

ITensorNetworks.CURRENT_PARTITIONING_BACKEND

ITensorNetworks.Backend

ITensorNetworks.ITensorNetwork

ITensorNetworks.TreeTensorNetworkState

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.current_partitioning_backend

ITensorNetworks.find_subgraph

ITensorNetworks.get_single_site_expec

ITensorNetworks.get_two_site_expec

ITensorNetworks.iterate_single_site_expec

ITensorNetworks.iterate_two_site_expec

ITensorNetworks.set_partitioning_backend!

ITensorNetworks.subgraphs

ITensorNetworks.two_site_rdm_bp

ITensorNetworks.update_all_mts

ITensorNetworks.update_mt

ITensorNetworks.CURRENT_PARTITIONING_BACKEND — Constant

Current default graph partitioning backend

ITensorNetworks.Backend — Type

Graph partitioning backend

ITensorNetworks.ITensorNetwork — Type

ITensorNetwork

ITensorNetworks.TreeTensorNetworkState — Type

TreeTensorNetworkState <: AbstractITensorNetwork

Fields

itensor_network::ITensorNetwork{V}

ortho_lims::Vector{V}: A vector of vertices defining the orthogonality limits.

ITensorNetworks.contraction_sequence — Method

GreedyMethod(; method=MinSpaceOut(), nrepeat=10)

The fast but poor greedy optimizer. Input arguments are:

method is MinSpaceDiff() or MinSpaceOut.

MinSpaceOut choose one of the contraction that produces a minimum output tensor size,

MinSpaceDiff choose one of the contraction that decrease the space most.

nrepeat is the number of repeatition, returns the best contraction order.

ITensorNetworks.contraction_sequence — Method

KaHyParBipartite(; sc_target, imbalances=collect(0.0:0.005:0.8),
                   max_group_size=40, greedy_config=GreedyMethod())

Optimize the einsum code contraction order using the KaHyPar + Greedy approach. This program first recursively cuts the tensors into several groups using KaHyPar, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are:

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

imbalances is a KaHyPar parameter that controls the group sizes in hierarchical bipartition,

max_group_size is the maximum size that allowed to used greedy search,

greedy_config is a greedy optimizer.

References

Hyper-optimized tensor network contraction

Simulating the Sycamore quantum supremacy circuits

ITensorNetworks.contraction_sequence — Method

SABipartite(; sc_target=25, ntrials=50, βs=0.1:0.2:15.0, niters=1000
              max_group_size=40, greedy_config=GreedyMethod(), initializer=:random)

Optimize the einsum code contraction order using the Simulated Annealing bipartition + Greedy approach. This program first recursively cuts the tensors into several groups using simulated annealing, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are:

size_dict, a dictionary that specifies leg dimensions,

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

max_group_size is the maximum size that allowed to used greedy search,

βs is a list of inverse temperature 1/T,

niters is the number of iteration in each temperature,

ntrials is the number of repetition (with different random seeds),

greedy_config configures the greedy method,

initializer, the partition configuration initializer, one can choose :random or :greedy (slow but better).

References

Hyper-optimized tensor network contraction

ITensorNetworks.contraction_sequence — Method

TreeSA(; sc_target=20, βs=collect(0.01:0.05:15), ntrials=10, niters=50,
         sc_weight=1.0, rw_weight=0.2, initializer=:greedy, greedy_config=GreedyMethod(; nrepeat=1))

Optimize the einsum contraction pattern using the simulated annealing on tensor expression tree.

sc_target is the target space complexity,

ntrials, βs and niters are annealing parameters, doing ntrials indepedent annealings, each has inverse tempteratures specified by βs, in each temperature, do niters updates of the tree.

sc_weight is the relative importance factor of space complexity in the loss compared with the time complexity.

rw_weight is the relative importance factor of memory read and write in the loss compared with the time complexity.

initializer specifies how to determine the initial configuration, it can be :greedy or :random. If it is using :greedy method to generate the initial configuration, it also uses two extra arguments greedy_method and greedy_nrepeat.

nslices is the number of sliced legs, default is 0.

fixed_slices is a vector of sliced legs, default is [].

References

Recursive Multi-Tensor Contraction for XEB Verification of Quantum Circuits

ITensorNetworks.current_partitioning_backend — Method

Get the graph partitioning backend

ITensorNetworks.find_subgraph — Method

Find the subgraph which contains the specified vertex.

TODO: Rename something more general, like:

findfirstinvertex_data(item, graph::AbstractDataGraph)

ITensorNetworks.get_single_site_expec — Method

given two flat networks psi and psi0, calculate the ratio of their contraction centred on the the subgraph containing v. The message tensors should be formulated over psi Link indices between psi and psi0 should be consistent so the mts can be applied to both

ITensorNetworks.get_two_site_expec — Method

given two flat networks psi and psi0, calculate the ratio of their contraction centred on the subgraph(s) containing v1 and v2. The message tensors should be formulated over psi.

ITensorNetworks.iterate_single_site_expec — Method

Starting with initial guess for messagetensors, monitor the convergence of an observable on a single site v (which is emedded in tnO)

ITensorNetworks.iterate_two_site_expec — Method

Starting with initial guess for messagetensors, monitor the convergence of an observable on a pair of sites v1 and v2 (which is emedded in tnO)

ITensorNetworks.set_partitioning_backend! — Method

Set the graph partitioning backend

ITensorNetworks.subgraphs — Method

Given a graph g, form 'nv(g) ÷ nverticesperpartition' subgraphs. Try to keep all subgraphs the same size and minimise edges cut between them Returns a datagraph where each vertex contains the list of vertices involved in that subgraph. The edges state which subgraphs are connected KaHyPar needs to be installed for this function to work

ITensorNetworks.two_site_rdm_bp — Method

Get twositerdm using belief propagation messagetensors

ITensorNetworks.update_all_mts — Method

Do an update of all message tensors for a given flat ITensornetwork and its partition into sub graphs

ITensorNetworks.update_mt — Method

DO a single update of a message tensor using the current subgraph and the incoming mts