ITensorNetworks

Documentation for ITensorNetworks.

ITensorNetworks.CURRENT_PARTITIONING_BACKEND

ITensorNetworks.Backend

ITensorNetworks.ITensorNetwork

ITensorNetworks.ProjTTN

ITensorNetworks.ProjTTNSum

ITensorNetworks.SweepStep

ITensorNetworks.TDVPInfo

ITensorNetworks.TTN

ITensorNetworks.TreeTensorNetwork

ITensorNetworks._binary_partition

ITensorNetworks._binary_tree_partition_inds

ITensorNetworks._distance

ITensorNetworks._introot_union!

ITensorNetworks._maxweightoutinds_tn

ITensorNetworks._mincut

ITensorNetworks._mincut_inds

ITensorNetworks._mincut_partitions

ITensorNetworks._mps_partition_inds_order

ITensorNetworks._remove_deltas

ITensorNetworks._root_union!

ITensorNetworks.alternating_update

ITensorNetworks.binary_tree_partition

ITensorNetworks.calculate_contraction

ITensorNetworks.compute_message_tensors

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence

ITensorNetworks.contraction_sequence_to_graph

ITensorNetworks.contraction_tree_leaf_bipartition

ITensorNetworks.current_partitioning_backend

ITensorNetworks.delta_network

ITensorNetworks.distance_from_roots

ITensorNetworks.distance_to_leaf

ITensorNetworks.fill_contraction_sequence_graph_vertices!

ITensorNetworks.find_subgraph

ITensorNetworks.findall_on_edges

ITensorNetworks.findall_on_vertices

ITensorNetworks.findfirst_on_edges

ITensorNetworks.findfirst_on_vertices

ITensorNetworks.finite_state_machine

ITensorNetworks.fsmTTN

ITensorNetworks.get_environment

ITensorNetworks.has_leaf_neighbor

ITensorNetworks.heisenberg

ITensorNetworks.heisenberg

ITensorNetworks.internal_edges

ITensorNetworks.is_leaf_edge

ITensorNetworks.ising

ITensorNetworks.ising

ITensorNetworks.ising_network

ITensorNetworks.ising_network

ITensorNetworks.leaf_vertices

ITensorNetworks.nested_graph_leaf_vertices

ITensorNetworks.partition

ITensorNetworks.partition_vertices

ITensorNetworks.partition_vertices

ITensorNetworks.randomITensorNetwork

ITensorNetworks.relabel_sites

ITensorNetworks.set_partitioning_backend!

ITensorNetworks.subgraphs

ITensorNetworks.subgraphs

ITensorNetworks.svdTTN

ITensorNetworks.tdvp

ITensorNetworks.to_vec

ITensorNetworks.update_all_mts

ITensorNetworks.update_mt

ITensors.apply

ITensors.dmrg

KrylovKit.eigsolve

KrylovKit.linsolve

NDTensors.contract

Observers.update!

ITensorNetworks.CURRENT_PARTITIONING_BACKEND — Constant

Current default graph partitioning backend

ITensorNetworks.Backend — Type

Graph partitioning backend

ITensorNetworks.ITensorNetwork — Type

ITensorNetwork

ITensorNetworks.ProjTTN — Type

ProjTTN

ITensorNetworks.ProjTTNSum — Type

ProjTTNSum

ITensorNetworks.SweepStep — Type

struct SweepStep{V}

Auxiliary object specifying a single local update step in a tree sweeping algorithm.

ITensorNetworks.TDVPInfo — Type

#fields

maxtruncerr::Float64: the maximum tuncation error

ITensorNetworks.TTN — Method

TTN(os::OpSum, sites::IndsNetwork{<:Index}; kwargs...)
TTN(eltype::Type{<:Number}, os::OpSum, sites::IndsNetwork{<:Index}; kwargs...)

Convert an OpSum object os to a TreeTensorNetwork, with indices given by sites.

ITensorNetworks.TreeTensorNetwork — Type

TreeTensorNetwork{V} <: AbstractTreeTensorNetwork{V}

Fields

itensor_network::ITensorNetwork{V}

ortho_lims::Vector{V}: A vector of vertices defining the orthogonality limits.

ITensorNetworks._binary_partition — Method

Partition the input network containing both tn and deltas (a vector of delta tensors) into two partitions, one adjacent to source_inds and the other adjacent to other external inds of the network.

ITensorNetworks._binary_tree_partition_inds — Method

Given a tn and outinds (a subset of noncommoninds of tn), get a binary tree structure of outinds that will be used in the binary tree partition. If maximally_unbalanced=true, the binary tree will have a line/mps structure. The binary tree is recursively constructed from leaves to the root.

Example:

TODO

ITensorNetworks._distance — Method

Sum of shortest path distances among all outinds.

ITensorNetworks._introot_union! — Method

Rewrite of the function DataStructures.root_union!(s::IntDisjointSet{T}, x::T, y::T) where {T<:Integer}.

ITensorNetworks._maxweightoutinds_tn — Method

create a tn with empty ITensors whose outinds weights are MAXWEIGHT The maxweighttn is constructed so that only commoninds of the tn will be considered in mincut.

ITensorNetworks._mincut — Method

Calculate the mincut between two subsets of the uncontracted inds (sourceinds and terminalinds) of the input tn. Mincut of two inds list is defined as the mincut of two newly added vertices, each one neighboring to one inds subset.

ITensorNetworks._mincut_inds — Method

Find a vector of indices within sourceindslist yielding the mincut of given tnpair. Args: tnpair: a pair of tns (tn1 => tn2), where tn2 is generated via _maxweightoutindstn(tn1) outtomaxweightind: a dict mapping each out ind in tn1 to out ind in tn2 sourceindslist: a list of vector of indices to be considered Note: For each sourceinds in sourceindslist, we consider its mincut within both tns (tn1, tn2) given in tnpair. The mincut in tn1 represents the rank upper bound when splitting sourceinds with other inds in outinds. The mincut in tn2 represents the rank upper bound when the weights of outinds are very large. The first mincut upper_bounds the number of non-zero singular values, while the second empirically reveals the singular value decay. We output the sourceinds where the first mincut value is the minimum, the secound mincut value is also the minimum under the condition that the first mincut is optimal, and the sourceinds have the lowest all-pair shortest path.

ITensorNetworks._mincut_partitions — Method

Calculate the mincutpartitions between two subsets of the uncontracted inds (sourceinds and terminal_inds) of the input tn.

ITensorNetworks._mps_partition_inds_order — Method

Given a tn and outinds, returns a vector of indices representing MPS inds ordering.

ITensorNetworks._remove_deltas — Method

Given an input tensor network containing tensors in the input tnand tensors indeltas`, remove redundent delta tensors indeltasand change inds accordingly to make the outputtnandoutdeltasrepresent the same tensor network but with less delta tensors. Note: inds of tensors intnanddeltasmay be changed, andoutdeltas` may still contain necessary delta tensors.

======== Example: julia> is = [Index(2, "i") for i in 1:6] julia> a = ITensor(is[1], is[2]) julia> b = ITensor(is[2], is[3]) julia> delta1 = delta(is[3], is[4]) julia> delta2 = delta(is[5], is[6]) julia> tn = ITensorNetwork([a,b]) julia> tn, outdeltas = ITensorNetworks.removedeltas(tn, [delta1, delta2]) julia> noncommoninds(Vector{ITensor}(tn)...) 2-element Vector{Index{Int64}}: (dim=2|id=339|"1") (dim=2|id=489|"4") julia> length(outdeltas) 1

ITensorNetworks._root_union! — Method

Rewrite of the function DataStructures.root_union!(s::DisjointSet{T}, x::T, y::T). The difference is that in the output of _root_union!, x is guaranteed to be the root of y when setting left_root=true, and y will be the root of x when setting left_root=false. In DataStructures.root_union!, the root value cannot be specified. A specified root is useful in functions such as _remove_deltas, where when we union two indices into one disjointset, we want the index that is the outinds if the given tensor network to always be the root in the DisjointSets.

ITensorNetworks.alternating_update — Method

tdvp(Hs::Vector{MPO},psi0::MPS,t::Number; kwargs...)
tdvp(Hs::Vector{MPO},psi0::MPS,t::Number, sweeps::Sweeps; kwargs...)

Use the time dependent variational principle (TDVP) algorithm to compute exp(t*H)*psi0 using an efficient algorithm based on alternating optimization of the MPS tensors and local Krylov exponentiation of H.

This version of tdvp accepts a representation of H as a Vector of MPOs, Hs = [H1,H2,H3,...] such that H is defined as H = H1+H2+H3+... Note that this sum of MPOs is not actually computed; rather the set of MPOs [H1,H2,H3,..] is efficiently looped over at each step of the algorithm when optimizing the MPS.

Returns:

psi::MPS - time-evolved MPS

ITensorNetworks.binary_tree_partition — Method

Given an input tn and a rooted binary tree of indices, return a partition of tn with the same binary tree structure as indsbtree. Note: in the output partition, we add multiple delta tensors to the network so that the output graph is guaranteed to be the same binary tree as indsbtree. Note: in the output partition, tensor vertex names will be changed. For a given input tensor with vertex name v, its name in the output partition will be(v, 1), and any delta tensor will have name(v, 2). Note: for a given binary tree with n indices, the output partition will contain 2n-1 vertices, with each leaf vertex corresponding to a sub tn adjacent to one output index. Keeping these leaf vertices in the partition makes laterapprox_itensornetwork` algorithms more efficient.

ITensorNetworks.calculate_contraction — Method

Calculate the contraction of a tensor network centred on the vertices verts. Using message tensors. Defaults to using tn[verts] as the local network but can be overriden

ITensorNetworks.compute_message_tensors — Method

Simulaneously initialise and update message tensors of a tensornetwork

ITensorNetworks.contraction_sequence — Method

GreedyMethod(; method=MinSpaceOut(), nrepeat=10)

The fast but poor greedy optimizer. Input arguments are:

method is MinSpaceDiff() or MinSpaceOut.

MinSpaceOut choose one of the contraction that produces a minimum output tensor size,

MinSpaceDiff choose one of the contraction that decrease the space most.

nrepeat is the number of repeatition, returns the best contraction order.

ITensorNetworks.contraction_sequence — Method

KaHyParBipartite(; sc_target, imbalances=collect(0.0:0.005:0.8),
                   max_group_size=40, greedy_config=GreedyMethod())

Optimize the einsum code contraction order using the KaHyPar + Greedy approach. This program first recursively cuts the tensors into several groups using KaHyPar, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are:

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

imbalances is a KaHyPar parameter that controls the group sizes in hierarchical bipartition,

max_group_size is the maximum size that allowed to used greedy search,

greedy_config is a greedy optimizer.

References

Hyper-optimized tensor network contraction

Simulating the Sycamore quantum supremacy circuits

ITensorNetworks.contraction_sequence — Method

SABipartite(; sc_target=25, ntrials=50, βs=0.1:0.2:15.0, niters=1000
              max_group_size=40, greedy_config=GreedyMethod(), initializer=:random)

Optimize the einsum code contraction order using the Simulated Annealing bipartition + Greedy approach. This program first recursively cuts the tensors into several groups using simulated annealing, with maximum group size specifed by max_group_size and maximum space complexity specified by sc_target, Then finds the contraction order inside each group with the greedy search algorithm. Other arguments are:

size_dict, a dictionary that specifies leg dimensions,

sc_target is the target space complexity, defined as log2(number of elements in the largest tensor),

max_group_size is the maximum size that allowed to used greedy search,

βs is a list of inverse temperature 1/T,

niters is the number of iteration in each temperature,

ntrials is the number of repetition (with different random seeds),

greedy_config configures the greedy method,

initializer, the partition configuration initializer, one can choose :random or :greedy (slow but better).

References

Hyper-optimized tensor network contraction

ITensorNetworks.contraction_sequence — Method

TreeSA(; sc_target=20, βs=collect(0.01:0.05:15), ntrials=10, niters=50,
         sc_weight=1.0, rw_weight=0.2, initializer=:greedy, greedy_config=GreedyMethod(; nrepeat=1))

Optimize the einsum contraction pattern using the simulated annealing on tensor expression tree.

sc_target is the target space complexity,

ntrials, βs and niters are annealing parameters, doing ntrials indepedent annealings, each has inverse tempteratures specified by βs, in each temperature, do niters updates of the tree.

sc_weight is the relative importance factor of space complexity in the loss compared with the time complexity.

rw_weight is the relative importance factor of memory read and write in the loss compared with the time complexity.

initializer specifies how to determine the initial configuration, it can be :greedy or :random. If it is using :greedy method to generate the initial configuration, it also uses two extra arguments greedy_method and greedy_nrepeat.

nslices is the number of sliced legs, default is 0.

fixed_slices is a vector of sliced legs, default is [].

References

Recursive Multi-Tensor Contraction for XEB Verification of Quantum Circuits

ITensorNetworks.contraction_sequence_to_graph — Method

Take a contractionsequence and return a graphical representation of it. The leaves of the graph represent the leaves of the sequence whilst the internalnodes of the graph define a tripartition of the graph and thus are named as an n = 3 element tuples, which each element specifying the keys involved. Edges connect parents/children within the contraction sequence.

ITensorNetworks.contraction_tree_leaf_bipartition — Method

Get the vertex bi-partition that a given edge represents

ITensorNetworks.current_partitioning_backend — Method

Get the graph partitioning backend

ITensorNetworks.delta_network — Method

RETURN A TENSOR NETWORK WITH COPY TENSORS ON EACH VERTEX. Note that passing a link_space will mean the indices of the resulting network don't match those of the input indsnetwork

ITensorNetworks.distance_from_roots — Method

Return all vertices which are within a certain pathlength dist of the leaves of the graph

ITensorNetworks.distance_to_leaf — Method

Get distance of a vertex from a leaf

ITensorNetworks.fill_contraction_sequence_graph_vertices! — Method

Given a contraction sequence which is a subsequence of some larger sequence (with leaves leaves) which is being built on g Spawn contract sequence' as a vertex oncurrent_g' and continue on with its children

ITensorNetworks.find_subgraph — Method

Find the subgraph which contains the specified vertex.

TODO: Rename something more general, like:

findfirstinvertex_data(item, graph::AbstractDataGraph)

ITensorNetworks.findall_on_edges — Method

Find all edges e such that f(graph[e]) == true

ITensorNetworks.findall_on_vertices — Method

Find all vertices v such that f(graph[v]) == true

ITensorNetworks.findfirst_on_edges — Method

Find the edge e such that f(graph[e]) == true

ITensorNetworks.findfirst_on_vertices — Method

Find the vertex v such that f(graph[v]) == true

ITensorNetworks.finite_state_machine — Method

finite_state_machine(os::OpSum{C}, sites::IndsNetwork{V,<:Index}, root_vertex::V) where {C,V}

Finite state machine generator for ITensors.OpSum Hamiltonian defined on a tree graph. The site Index graph must be a tree graph, and the chosen root must be a leaf vertex of this tree. Returns a DataGraph of SparseArrayKit.SparseArrays

ITensorNetworks.fsmTTN — Method

fsmTTN(os::OpSum{C}, sites::IndsNetwork{V,<:Index}, root_vertex::V, kwargs...) where {C,V}

Construct a dense TreeTensorNetwork from sparse finite state machine represenatation, without compression.

ITensorNetworks.get_environment — Method

Given a subet of vertices of a given Tensor Network and the Message Tensors for that network, return a Dictionary with the involved subgraphs as keys and the vector of tensors associated with that subgraph as values Specifically, the contraction of the environment tensors and tn[vertices] will be a scalar.

ITensorNetworks.has_leaf_neighbor — Method

Determine if a node has any neighbors which are leaves

ITensorNetworks.heisenberg — Method

Random field J1-J2 Heisenberg model on a general graph

ITensorNetworks.heisenberg — Method

Random field J1-J2 Heisenberg model on a chain of length N

ITensorNetworks.internal_edges — Method

Get all edges which do not involve a leaf

ITensorNetworks.is_leaf_edge — Method

Determine if an edge involves a leaf (at src or dst)

ITensorNetworks.ising — Method

Next-to-nearest-neighbor Ising model (ZZX) on a general graph

ITensorNetworks.ising — Method

Next-to-nearest-neighbor Ising model (ZZX) on a chain of length N

ITensorNetworks.ising_network — Method

BUILD Z OF CLASSICAL ISING MODEL ON A GIVEN GRAPH AT INVERSE TEMP BETA H = -\sum{(v,v') \in edges}\sigma^{z}{v}\sigma^{z}_{v'} OPTIONAL ARGUMENT: h: EXTERNAL MAGNETIC FIELD szverts: A LIST OF VERTICES OVER WHICH TO APPLY A SZ. THE RESULTANT NETWORK CAN THEN BE CONTRACTED AND DIVIDED BY THE ACTUAL PARTITION FUNCTION TO GET THAT OBSERVABLE INDSNETWORK IS ASSUMED TO BE BUILT FROM A GRAPH (NO SITE INDS) AND OF LINK SPACE 2

ITensorNetworks.ising_network — Method

BUILD Z OF CLASSICAL ISING MODEL ON A GIVEN GRAPH AT INVERSE TEMP BETA H = -\sum{(v,v') \in edges}\sigma^{z}{v}\sigma^{z}_{v'} TAKE AS AN OPTIONAL ARGUMENT A LIST OF VERTICES OVER WHICH TO APPLY A SZ. THE RESULTANT NETWORK CAN THEN BE CONTRACTED AND DIVIDED BY THE ACTUAL PARTITION FUNCTION TO GET THAT OBSERVABLE

ITensorNetworks.leaf_vertices — Method

Get all vertices which are leaves of a graph

ITensorNetworks.nested_graph_leaf_vertices — Method

Given a nested graph fetch all the vertices down to the lowest levels and return the grouping at the highest level. Keyword argument is used to state whether we are at the top

ITensorNetworks.partition — Method

partition(g::AbstractGraph; npartitions::Integer, kwargs...)
partition(g::AbstractGraph, subgraph_vertices)

Given a graph g, partition g into npartitions partitions or into partitions with nvertices_per_partition vertices per partition. The partitioning tries to keep all subgraphs the same size and minimize edges cut between them.

Alternatively, specify a desired partitioning with a collection of sugraph vertices.

Returns a data graph where each vertex contains the corresponding subgraph as vertex data. The edges indicates which subgraphs are connected, and the edge data stores a dictionary with two fields. The field :edges stores a list of the edges of the original graph that were connecting the two subgraphs, and :edge_data stores a dictionary mapping edges of the original graph to the data living on the edges of the original graph, if it existed.

Therefore, one should be able to extract that data and recreate the original graph from the results of partition.

A graph partitioning backend such as Metis or KaHyPar needs to be installed for this function to work if the subgraph vertices aren't specified explicitly.

ITensorNetworks.partition_vertices — Method

partition_vertices(g::AbstractGraph, subgraph_vertices::Vector)

Given a graph (g) and groups of vertices defining subgraphs of that graph (subgraph_vertices), return a DataGraph storing the subgraph vertices on the vertices of the graph and with edges denoting which subgraphs of the original graph have edges connecting them, along with edge data storing the original edges that were connecting the subgraphs.

ITensorNetworks.partition_vertices — Method

partition_vertices(g::AbstractGraph; npartitions, nvertices_per_partition, kwargs...)

Given a graph g, partition the vertices of g into 'npartitions' partitions or into partitions with nvertices_per_partition vertices per partition. Try to keep all subgraphs the same size and minimise edges cut between them Returns a datagraph where each vertex contains the list of vertices involved in that subgraph. The edges state which subgraphs are connected. A graph partitioning backend such as Metis or KaHyPar needs to be installed for this function to work.

ITensorNetworks.randomITensorNetwork — Method

Build an ITensor network on a graph specified by the inds network s. Bonddim is given by linkspace and entries are randomised (normal distribution, mean 0 std 1)

ITensorNetworks.relabel_sites — Method

Relabel sites in OpSum according to given site map

ITensorNetworks.set_partitioning_backend! — Method

Set the graph partitioning backend

ITensorNetworks.subgraphs — Method

subgraphs(g::AbstractGraph, subgraph_vertices)

Return a collection of subgraphs of g defined by the subgraph vertices subgraph_vertices.

ITensorNetworks.subgraphs — Method

subgraphs(g::AbstractGraph; npartitions::Integer, kwargs...)

Given a graph g, partition g into npartitions partitions or into partitions with nvertices_per_partition vertices per partition, returning a list of subgraphs. Try to keep all subgraphs the same size and minimise edges cut between them. A graph partitioning backend such as Metis or KaHyPar needs to be installed for this function to work.

ITensorNetworks.svdTTN — Method

svdTTN(os::OpSum{C}, sites::IndsNetwork{V<:Index}, root_vertex::V, kwargs...) where {C,V}

Construct a dense TreeTensorNetwork from a symbolic OpSum representation of a Hamiltonian, compressing shared interaction channels.

ITensorNetworks.tdvp — Method

tdvp(H::TTN, t::Number, psi0::TTN; kwargs...)

Use the time dependent variational principle (TDVP) algorithm to compute exp(H*t)*psi0 using an efficient algorithm based on alternating optimization of the state tensors and local Krylov exponentiation of H.

Returns:

psi - time-evolved state

Optional keyword arguments:

outputlevel::Int = 1 - larger outputlevel values resulting in printing more information and 0 means no output

observer - object implementing the Observer interface which can perform measurements and stop early

write_when_maxdim_exceeds::Int - when the allowed maxdim exceeds this value, begin saving tensors to disk to free memory in large calculations

ITensorNetworks.to_vec — Method

to_vec(x)

Transform x into a Vector. Returns the vector and a closure which inverts the transformation.

Modeled after FiniteDifferences.to_vec:

https://github.com/JuliaDiff/FiniteDifferences.jl/blob/main/src/to_vec.jl

ITensorNetworks.update_all_mts — Method

Do an update of all message tensors for a given ITensornetwork and its partition into sub graphs

ITensorNetworks.update_mt — Method

DO a single update of a message tensor using the current subgraph and the incoming mts

ITensors.apply — Method

Overload of ITensors.apply.

ITensors.dmrg — Method

Overload of ITensors.dmrg.

KrylovKit.eigsolve — Method

Overload of KrylovKit.eigsolve.

KrylovKit.linsolve — Function

linsolve(
    A::ITensorNetworks.AbstractTreeTensorNetwork,
    b::ITensorNetworks.AbstractTreeTensorNetwork,
    x₀::ITensorNetworks.AbstractTreeTensorNetwork
)
linsolve(
    A::ITensorNetworks.AbstractTreeTensorNetwork,
    b::ITensorNetworks.AbstractTreeTensorNetwork,
    x₀::ITensorNetworks.AbstractTreeTensorNetwork,
    a₀::Number
)
linsolve(
    A::ITensorNetworks.AbstractTreeTensorNetwork,
    b::ITensorNetworks.AbstractTreeTensorNetwork,
    x₀::ITensorNetworks.AbstractTreeTensorNetwork,
    a₀::Number,
    a₁::Number;
    kwargs...
)

Compute a solution x to the linear system:

(a₀ + a₁ * A)*x = b

using starting guess x₀. Leaving a₀, a₁ set to their default values solves the system A*x = b.

To adjust the balance between accuracy of solution and speed of the algorithm, it is recommed to first try adjusting the solver_tol keyword argument descibed below.

Keyword arguments:

ishermitian::Bool=false - should set to true if the MPO A is Hermitian

solver_krylovdim::Int=30 - max number of Krylov vectors to build on each solver iteration

solver_maxiter::Int=100 - max number outer iterations (restarts) to do in the solver step

solver_tol::Float64=1E-14 - tolerance or error goal of the solver

Overload of KrylovKit.linsolve.

NDTensors.contract — Method

Overload of ITensors.contract.

Observers.update! — Method

Overload of Observers.update!.